updated the C# client to get it compiling and most of the tests working with the latest generated OpenWire code; particularly after the change of Command.CommandId from short to int
git-svn-id: https://svn.apache.org/repos/asf/incubator/activemq/trunk@386440 13f79535-47bb-0310-9956-ffa450edef68

+++ activemq-dotnet/src/main/csharp/ActiveMQ/Commands/Command.cs
@@ -18,13 +18,13 @@ using ActiveMQ.Commands;
 
 namespace ActiveMQ.Commands
 {
-	
-	/// <summary>
-	/// An OpenWire command
-	/// </summary>
-	public interface Command : DataStructure
+    
+    /// <summary>
+    /// An OpenWire command
+    /// </summary>
+    public interface Command : DataStructure
     {
-        short CommandId
+        int CommandId
         {
             get;
             set;
